refactor: 결제 페이지 구조 변경 및 결제 수단 선택 기능 추가
